Support nullable element types in collections.
A number of refactoring was done in the process:

1) SchemaOption is now const constructable. There is no longer a
need to register a factory, as the factory is known at compile time.

2) RealmSchema supports lookup of a SchemaOption by type.

3) In a similar vain, properties can be looked-up by key on
RealmMetadata efficiently.

3) Generator will name all const constructed schema properties (which
are now a generic type called ValueProperty<T>).

4) Getter and setters on realm objects will trampoline through the
associated ValueProperty, and there are special ListProperty and
ObjectProperty subclasses, that allow compile type access to the link
target type.

5) RealmAccessor now has separate methods for reading objects, lists,
and plain properties.